# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ParcelGuard API
  description: |
    REST API for ParcelGuard - a DIY multi-camera security system for monitoring
    communal areas and protecting deliveries.
  version: 1.0.0
  contact:
    name: ParcelGuard
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token obtained from login
  schemas:
...
    MotionEvent:
      type: object
      properties:
        id:
          type: string
        cameraId:
          type: string
        timestamp:
          type: integer
          description: Unix timestamp in seconds
        duration:
          type: integer
          nullable: true
          description: Duration in seconds
        thumbnailPath:
          type: string
          nullable: true
        videoPath:
          type: string
          nullable: true
        isImportant:
          type: boolean
        isFalseAlarm:
          type: boolean
        createdAt:
          type: integer
...
security:
  - bearerAuth: []
paths:
...
  /events:
    get:
      tags: [Events]
      summary: List events with filtering
      parameters:
        - name: cameraId
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: integer
        - name: endDate
          in: query
          schema:
            type: integer
        - name: isImportant
          in: query
          schema:
            type: boolean
        - name: isFalseAlarm
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Paginated events
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      events:
                        type: array
                        items:
                          $ref: '#/components/schemas/MotionEvent'
                      total:
                        type: integer
                      page:
                        type: integer
                      pageSize:
                        type: integer
                      hasMore:
                        type: boolean
...
  /events/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Events]
      summary: Get event by ID
      responses:
        '200':
          description: Event details
        '404':
          description: Event not found
    put:
      tags: [Events]
      summary: Update event flags
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                isImportant:
                  type: boolean
                isFalseAlarm:
                  type: boolean
      responses:
        '200':
          description: Event updated
        '404':
          description: Event not found
    delete:
      tags: [Events]
      summary: Delete event
      responses:
        '200':
          description: Event deleted
        '404':
          description: Event not found
  /events/bulk-delete:
    post:
      tags: [Events]
      summary: Delete multiple events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Events deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      deletedCount:
                        type: integer
  /events/{id}/thumbnail:
    get:
      tags: [Events]
      summary: Get event thumbnail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Thumbnail image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        '404':
          description: Not found
  /events/{id}/video:
    get:
      tags: [Events]
      summary: Stream event video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Video stream
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        '404':
          description: Not found
  /events/{id}/download:
    get:
      tags: [Events]
      summary: Download event video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Video file download
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        '404':
          description: Not found
...
  /motion/events:
    post:
      tags: [Events]
      summary: Motion daemon webhook endpoint
      description: Receives motion events from Motion daemon. Called internally.
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [cameraId, eventId, type, timestamp]
              properties:
                cameraId:
                  type: string
                  description: Camera identifier
                eventId:
                  type: string
                  description: Unique event identifier from Motion
                type:
                  type: string
                  enum: [start, end]
                  description: Event type
                timestamp:
                  type: integer
                  description: Unix timestamp (seconds)
      responses:
        '201':
          description: Event created (on start)
        '200':
          description: Event updated (on end)
        '400':
          description: Invalid payload